requirements.txt
streamlit==1.44.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.3.2
plotly==5.18.0
seaborn==0.13.2
pillow==10.2.0